genx-ai/requirements.txt
fastapi
uvicorn
requests
huggingface_hub
nltk
pyspellchecker
Pillow
python-multipart
opencv-python
git+https://github.com/Harsh-BH/Real-ESRGAN.git
torch
aiohttp